app/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
databricks-sdk>=0.20.0
openai>=1.10.0
